Split the web adapter's socket suite into web_integration.rs
The web adapter was the only one whose service-backed tests lived in the
tier-1 file. tests/web_adapter.rs was gated #![cfg(feature = "web")] and
held the full WebSocket suite plus the TLS round trip, so CI's `test` job
ran all of it: that job is

  cargo nextest --all-features --lib --tests -E 'not binary(/_integration$/)'

and web_adapter does not match the exclusion, while --all-features turns
on web-tls-integration-test. Thirteen socket tests therefore ran twice —
once in the fast job and again in web-integration.yml — which is why the
web tests have been a recurring source of timeouts and flakes (#691,
77f9744).

The filename suffix is the entire mechanism, so the fix is the filename:

- tests/web_integration.rs (new, #![cfg(feature = "web-integration-test")])
  takes the 8 tests that connect a real client, plus the client helpers.
- tests/web_adapter.rs keeps the 5 that need nothing listening: the two
  bind cases, the two historical contracts, and the TLS-config error.
  It now runs in 0.00s.
- web-integration-test = ["web"] is a new no-dep feature, matching how
  zmq/fix/kdb gate their socket-only tiers. web-tls-integration-test
  implies it, so the single flag the workflow passes still compiles and
  runs everything.

web-integration.yml pinned --test web_adapter. Left as it was, the 8
moved tests would have run nowhere at all while the job stayed green, so
it now names both binaries and its paths: trigger includes the new file.

Verified: tier 1 4/4 in 0.00s; tier 2 8/8 including TLS in 4.35s; and
`cargo nextest list` under CI's exact filter now returns only the 5
web_adapter tests, with all 13 present when the filter is dropped.

web/CLAUDE.md claimed "There is no separate web_integration.rs" because
the round trips need no service. That rationale was the wrong axis — the
split is about speed and load-sensitivity, not about standing up a
service — so the Tests section now says which axis it is and warns that a
socket test added to web_adapter.rs silently rejoins the fast job.
